New comments and a few aesthetic changes
We are back into this! My modified/new plan with this is to decouple the planet, star and planetary system generator. However, before that I want to finish going in the direction I am currently going in, and get a V0.1 release or something. Then look into decoupling when the logic is atleast in place.